Rename global to globaldb
