+++ cmd/portal/server/wire_gen.go
@@ -8,7 +8,7 @@ package server
 import (
 	"context"
 	"github.com/authgear/authgear-server/pkg/lib/config/configsource"
-	"github.com/authgear/authgear-server/pkg/lib/infra/db/global"
+	"github.com/authgear/authgear-server/pkg/lib/infra/db/globaldb"
 	"github.com/authgear/authgear-server/pkg/portal/deps"
 	"github.com/authgear/authgear-server/pkg/util/clock"
 )
@@ -30,10 +30,10 @@ func newConfigSourceController(p *deps.RootProvider, c context.Context) *configs
 	trustProxy := environmentConfig.TrustProxy
 	clock := _wireSystemClockValue
 	databaseEnvironmentConfig := &environmentConfig.Database
-	sqlBuilder := global.NewSQLBuilder(databaseEnvironmentConfig)
+	sqlBuilder := globaldb.NewSQLBuilder(databaseEnvironmentConfig)
 	pool := p.Database
-	handle := global.NewHandle(c, pool, databaseEnvironmentConfig, factory)
-	sqlExecutor := global.NewSQLExecutor(c, handle)
+	handle := globaldb.NewHandle(c, pool, databaseEnvironmentConfig, factory)
+	sqlExecutor := globaldb.NewSQLExecutor(c, handle)
 	store := &configsource.Store{
 		SQLBuilder:  sqlBuilder,
 		SQLExecutor: sqlExecutor,
