+++ pkg/lib/infra/db/globaldb/deps.go
@@ -1,4 +1,4 @@
-package global
+package globaldb
 
 import (
 	"context"
