+++ pkg/portal/deps.go
@@ -4,7 +4,7 @@ import (
 	"github.com/google/wire"
 
 	adminauthz "github.com/authgear/authgear-server/pkg/lib/admin/authz"
-	globaldb "github.com/authgear/authgear-server/pkg/lib/infra/db/global"
+	"github.com/authgear/authgear-server/pkg/lib/infra/db/globaldb"
 	"github.com/authgear/authgear-server/pkg/portal/deps"
 	"github.com/authgear/authgear-server/pkg/portal/endpoint"
 	"github.com/authgear/authgear-server/pkg/portal/graphql"
