+++ pkg/portal/transport/admin_api_handler.go
@@ -6,7 +6,7 @@ import (
 
 	relay "github.com/authgear/graphql-go-relay"
 
-	globaldb "github.com/authgear/authgear-server/pkg/lib/infra/db/global"
+	"github.com/authgear/authgear-server/pkg/lib/infra/db/globaldb"
 	"github.com/authgear/authgear-server/pkg/portal/session"
 	"github.com/authgear/authgear-server/pkg/util/httproute"
 	"github.com/authgear/authgear-server/pkg/util/log"
